requirements.txt
python-telegram-bot[job-queue]>=21.9
requests==2.31.0
beautifulsoup4==4.12.3
lxml>=5.3.0